+++ test-d/delimiter-cased-properties-deep.ts
@@ -1,5 +1,5 @@
 import {expectType} from 'tsd';
-import {DelimiterCasedPropertiesDeep} from '../ts41/delimiter-cased-props-deep';
+import {DelimiterCasedPropertiesDeep} from '../ts41/delimiter-cased-properties-deep';
 
 declare const foo: DelimiterCasedPropertiesDeep<{helloWorld: {fooBar: string}}, '/'>;
 expectType<{'hello/world': {'foo/bar': string}}>(foo);
