+++ ts41/camel-cased-properties-deep.d.ts
