+++ ts41/camel-cased-properties.d.ts
