+++ ts41/delimiter-cased-properties-deep.d.ts
