+++ ts41/delimiter-cased-properties.d.ts
