+++ ts41/pascal-cased-properties-deep.d.ts
