+++ ts41/pascal-cased-properties.d.ts
