+++ ts41/snake-cased-properties.d.ts
@@ -1,4 +1,4 @@
-import {DelimiterCasedProperties} from './delimiter-cased-props';
+import {DelimiterCasedProperties} from './delimiter-cased-properties';
 
 /**
 Convert object props to snake_case but not recursively.
